requirements.txt
SQLAlchemy
numpy
pandas
pandas-stubs
psycopg2-binary
pydantic
python-dotenv
sqlacodegen
